move service's tests into service module, due using mock clock
